+++ utility.lua
@@ -1,0 +1,37 @@
+function initWorld()
+	-- world
+	world = bump.newWorld()
+	-- entity table
+	chars = {}
+	-- castles
+	heroCastle = {x=heroCastleX, y=castleY, width=castleWidth, height=castleHeight, life=lifeCastle, speed=speedCastle, party="hero", charType="castle"}
+	enemyCastle = {x=enemyCastleX, y=castleY, width=castleWidth, height=castleHeight, life=lifeCastle, speed=speedCastle, party="enemy", charType="castle"}
+	addChar(heroCastle)
+	addChar(enemyCastle)
+	timeTillDeath = 0
+	counter = 0
+end
+
+function inScreen(x, y)
+	-- checks if coordinates are in the screen
+	-- minumum
+	local x = x
+	local y = y
+	if (x < 0) or (y < 0) then
+		return false
+	else return true
+	end
+	-- maximum
+	if (x < screenW) or (y < screenH) then
+		return false
+	else
+		return true
+	end
+
+end
+
+function removeWorld()
+	-- delete all chars in the world
+	chars = nil
+	char = nil
+end
